requirements.txt
pyCUDA
SHTns
DUCC